refactor: move all type definitions into a separate file
