+++ src/plugins/trackerModule.ts
@@ -1,45 +1,12 @@
 import firebase from "firebase";
-/* TODO: do something with user ids */
-export interface ITracker {
-  currentEvent: IEvent | null;
-  addUserInfo(dataObj: IEventData): void;
-  setEvent(strategy: IEvent): void;
-  events: {
-    [key in EventTypes]: IEvent;
-  };
-  track(eventType: string, data: IEventData): void;
-}
-
-export interface ITrackConfig {
-  userInfo: {
-    name?: string;
-    email: string;
-    uid: string;
-  };
-}
-
-export type ButtonClickData = {
-  [key in string | number]: any;
-} & {
-  eventName: string;
-};
-
-export type RouteChangeData = {
-  route: string;
-  enteredFrom: string;
-  exceptions?: Array<string>;
-};
-
-export type IEventData = ButtonClickData | RouteChangeData;
-
-interface IEvent {
-  handle(data: IEventData): void;
-}
-
-enum EventTypes {
-  ROUTE_CHANGE = "ROUTE_CHANGE",
-  BUTTON_CLICK = "BUTTON_CLICK",
-}
+import {
+  ITrackConfig,
+  ITracker,
+  EventTypes,
+  RouteChangeData,
+  ButtonClickData,
+  IEventData,
+} from "./trackerTypes";
 
 export function createTrackerModule(config: ITrackConfig): ITracker {
   return {
